Moved Machine into Domain and corrected 2 methods
