+++ T-800/Program.cs
@@ -6,7 +6,13 @@ namespace T_800
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
+            var t800 = new Machine("T800");
+
+            t800.ActivateMachine(true);
+            t800.DeactivateMachine(true);
+            
+            
+            
         }
     }
 }
